skiller-simulator-lookup: initial skeleton
This plugin should estimate skill execution times based on entries of a
mongodb collections. The idea is to feed a database with execution times
from multiple real and/or gazebo-simulated runs. This data provides a
sample space of each skills from which a random sample may be drawn
randomly.

This way the skiller simulator may be fed with realistic execution
times.

The structure of this skeleton is adapted from the
skiller-simulator-navgraph plugin.