skiller-simulator-lookup: add try-by-default option
It may be convenient to use this estimator for all skills unless
specified otherwise. Therefore, a new configurable flag is introduced
that inverts the semantics of the list of skills that may be estimated,
if it becomes true. So depending on this set the skill list is
interpreted as those skills that should be estimated by this plugin if
possible, or as those skills that should never be estimated.

A use case could be to estimate all skills except the navigational ones
via lookup, leaving skills such as goto to the navgraph based estimator.