mongodb_log: add thread id when logging blackboard
The thread id may be useful to distinguish between different writers of
the same interface, e.g., if multiple fawkes instances are running on
the same host in local multi-agent simulations.

In the context of the skiller-simulator lookup this becomes useful to
gather Skiller data from a gazebo simulation. The blackboard logger may
be used to track the physically simulated execution times that then form
the dataset of a skiller-simulator based simulation. In order to use the
blackboard log data it is important to distinguish between different
writers so that the data of multiple agents that perform the same skills
in the same time are not confused with each other.